// client-side/freelancer-platform/src/app/modules/chat/models/chat.model.ts
export class Chat {
    id: string = '';
    jobId: string = '';
    jobTitle: string = '';
    proposalId: string = '';
    contractId: string | null = null;
    clientId: string = '';
    isClientActive: boolean = false;
    clientName: string = '';
    freelancerId: string = '';
    isFreelancerActive: boolean = false;
    freelancerName: string = '';
}